feat: add createNutrition utility function
