+++ src/test/kotlin/logic/usecases/GetSeafoodByProteinContentUseCaseTest.kt
@@ -153,20 +153,3 @@ class GetSeafoodByProteinContentUseCaseTest {
         Truth.assertThat(ids).isEqualTo(listOf(1))
     }
 }
-fun createNutrition(
-    calories: Double? = null,
-    totalFat: Double? = null,
-    sugar: Double? = null,
-    sodium: Double? = null,
-    protein: Double? = null,
-    saturatedFat: Double? = null,
-    carbohydrates: Double? = null,
-) = Nutrition(
-    calories = calories,
-    totalFat = totalFat,
-    sugar = sugar,
-    sodium = sodium,
-    protein = protein,
-    saturatedFat = saturatedFat,
-    carbohydrates = carbohydrates,
-)
